Only swizzles sendEvent if config contains the touchesBegan trigger
Removes semaphore from UIWindow because all UIWindow events will be performed in serial as they're on the main actor.
